[feature/mypage]: refactor for sprint no.2
