--- app/src/main/java/com/teampophory/pophory/feature/home/mypage/util/GridSpacingCustomDecoration.kt
@@ -1,34 +1,0 @@
-package com.teampophory.pophory.feature.home.mypage.util
-
-import android.graphics.Rect
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-class GridSpacingCustomDecoration(private val spanCount: Int, private val spacing: Int, private val includeEdge: Boolean) : RecyclerView.ItemDecoration() {
-
-    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
-        val position = parent.getChildAdapterPosition(view) -1 // item position
-        val column = position % spanCount // item column
-
-        // If it's the first item (position 0), skip decoration
-        if (position == -1) {
-            return
-        }
-
-        if (!includeEdge) {
-            outRect.left = spacing - column * spacing / spanCount // spacing - column * ((1f / spanCount) * spacing)
-            outRect.right = (column + 1) * spacing / spanCount // (column + 1) * ((1f / spanCount) * spacing)
-
-            if (position < spanCount) { // top edge
-                outRect.top = spacing
-            }
-            outRect.bottom = spacing // item bottom
-        } else {
-            outRect.left = column * spacing / spanCount // column * ((1f / spanCount) * spacing)
-            outRect.right = spacing - (column + 1) * spacing / spanCount // spacing - (column + 1) * ((1f / spanCount) * spacing)
-            if (position >= spanCount) {
-                outRect.top = spacing // item top
-            }
-        }
-    }
-}
