+++ app/src/main/res/layout/fragment_mypage.xml
@@ -98,7 +98,9 @@
         android:layout_marginHorizontal="20dp"
         android:layout_marginTop="20dp"
         android:background="@drawable/bg_round_20dp_lightpurple"
-        android:padding="20dp"
+        android:paddingHorizontal="20dp"
+        android:paddingTop="20dp"
+        android:paddingBottom="18dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/iv_mypage_profile">
@@ -107,7 +109,6 @@
             android:id="@+id/tv_mypage_share_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="4dp"
             android:text="네컷사진 공유하기"
             android:textAppearance="?textAppearanceHeadlineMedium"
             android:textColor="?colorOnSurface100"
@@ -119,7 +120,6 @@
             android:id="@+id/iv_mypage_share_ic"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="4dp"
             android:src="@drawable/ic_chevron_right"
             app:layout_constraintStart_toEndOf="@id/tv_mypage_share_title"
             app:layout_constraintTop_toTopOf="parent" />
@@ -156,8 +156,10 @@
         android:layout_height="wrap_content"
         android:background="@drawable/bg_round_20dp_lightpurple"
         android:layout_marginHorizontal="20dp"
-        android:layout_marginTop="12dp"
-        android:padding="20dp"
+        android:layout_marginTop="14dp"
+        android:paddingHorizontal="20dp"
+        android:paddingTop="20dp"
+        android:paddingBottom="16dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/layout_mypage_share">
@@ -166,7 +168,6 @@
             android:id="@+id/tv_mypage_story_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="4dp"
             android:text="포릿 이야기 들으러 가기"
             android:textAppearance="?textAppearanceHeadlineMedium"
             android:textColor="?colorOnSurface100"
@@ -178,7 +179,6 @@
             android:id="@+id/iv_mypage_story_ic"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="4dp"
             android:src="@drawable/ic_chevron_right"
             app:layout_constraintStart_toEndOf="@id/tv_mypage_story_title"
             app:layout_constraintTop_toTopOf="parent" />
